Add labels for targets.
