+++ crates/workspace/Cargo.toml
@@ -9,6 +9,7 @@ moon_config = { path = "../config" }
 moon_error = { path = "../error" }
 moon_logger = { path = "../logger" }
 moon_project = { path = "../project" }
+moon_terminal = { path = "../terminal" }
 moon_toolchain = { path = "../toolchain" }
 moon_utils = { path = "../utils" }
 async-trait = "0.1"
